#409 Move custom unit definitions
