+++ neclib/core/__init__.py
@@ -2,6 +2,7 @@
 
 from . import formatting  # noqa: F401
 from . import type_aliases  # noqa: F401
+from . import units  # noqa: F401
 from .data_type import *  # noqa: F401, F403
 from .exceptions import *  # noqa: F401, F403
 from .files import *  # noqa: F401, F403
