+++ neclib/devices/signal_generator/fsw0010.py
@@ -4,7 +4,7 @@ from typing import Optional, Union
 import astropy.units as u
 import ogameasure
 
-from ...units import dBm
+from ...core.units import dBm
 from ...utils import busy, skip_on_simulator
 from .signal_generator_base import SignalGenerator
